function/requirements.txt
gpt4all==2.2.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.28
packaging
requests
loguru
PyYAML
python-telegram-bot==20.8